refactor/user : user adapter 삭제 및 관련 클래스 코드 수정
