+++ src/main/java/com/sparta/springtrello/auth/UserDetailsServiceImpl.java
@@ -1,9 +1,7 @@
 package com.sparta.springtrello.auth;
 
-
-
 import com.sparta.springtrello.domain.user.entity.User;
-import com.sparta.springtrello.domain.user.repository.UserAdapter;
+import com.sparta.springtrello.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -13,11 +11,14 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class UserDetailsServiceImpl implements UserDetailsService {
-    private final UserAdapter userAdapter;
+    private final UserService userService;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userAdapter.findByUsername(username);
+        User user = userService.findByUsername(username);
+        if (user == null) {
+            throw new UsernameNotFoundException("User not found with username: " + username);
+        }
         return new UserDetailsImpl(user);
     }
 }
