+++ src/main/java/com/sparta/springtrello/domain/user/repository/UserRepository.java
@@ -1,8 +1,5 @@
 package com.sparta.springtrello.domain.user.repository;
 
-
-
-
 import com.sparta.springtrello.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
