+++ src/main/java/com/sparta/springtrello/domain/user/service/KakaoService.java
@@ -6,7 +6,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.springtrello.auth.JwtProvider;
 import com.sparta.springtrello.domain.user.dto.KakaoUserInfoDto;
 import com.sparta.springtrello.domain.user.entity.User;
-import com.sparta.springtrello.domain.user.repository.UserAdapter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -26,7 +25,7 @@ import java.net.URI;
 @RequiredArgsConstructor
 public class KakaoService {
 
-    private final UserAdapter userAdapter;
+    private final UserService userService;
     private final RestTemplate restTemplate;
     private final JwtProvider jwtProvider;
 
@@ -37,20 +36,16 @@ public class KakaoService {
     private String redirectUri;
 
     public String kakaoLogin(String code) throws JsonProcessingException {
-        // 1. "인가 코드"로 "액세스 토큰" 요청
         String accessToken = getToken(code);
         log.info("Kakao access token: {}", accessToken);
 
-        // 2. 토큰으로 카카오 API 호출 : "액세스 토큰"으로 "카카오 사용자 정보" 가져오기
         KakaoUserInfoDto kakaoUserInfo = getKakaoUserInfo(accessToken);
         log.info("Kakao user info: {}", kakaoUserInfo);
 
-        // 3. JWT 토큰 생성
         return createJwtToken(kakaoUserInfo);
     }
 
     private String getToken(String code) throws JsonProcessingException {
-        // 요청 URL 만들기
         URI uri = UriComponentsBuilder
                 .fromUriString("https://kauth.kakao.com")
                 .path("/oauth/token")
@@ -58,11 +53,9 @@ public class KakaoService {
                 .build()
                 .toUri();
 
-        // HTTP Header 생성
         HttpHeaders headers = new HttpHeaders();
         headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
 
-        // HTTP Body 생성
         MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
         body.add("grant_type", "authorization_code");
         body.add("client_id", clientId);
@@ -74,19 +67,16 @@ public class KakaoService {
                 .headers(headers)
                 .body(body);
 
-        // HTTP 요청 보내기
         ResponseEntity<String> response = restTemplate.exchange(
                 requestEntity,
                 String.class
         );
 
-        // HTTP 응답 (JSON) -> 액세스 토큰 파싱
         JsonNode jsonNode = new ObjectMapper().readTree(response.getBody());
         return jsonNode.get("access_token").asText();
     }
 
     private KakaoUserInfoDto getKakaoUserInfo(String accessToken) throws JsonProcessingException {
-        // 요청 URL 만들기
         URI uri = UriComponentsBuilder
                 .fromUriString("https://kapi.kakao.com")
                 .path("/v2/user/me")
@@ -94,7 +84,6 @@ public class KakaoService {
                 .build()
                 .toUri();
 
-        // HTTP Header 생성
         HttpHeaders headers = new HttpHeaders();
         headers.add("Authorization", "Bearer " + accessToken);
         headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
@@ -104,7 +93,6 @@ public class KakaoService {
                 .headers(headers)
                 .body(new LinkedMultiValueMap<>());
 
-        // HTTP 요청 보내기
         ResponseEntity<String> response = restTemplate.exchange(
                 requestEntity,
                 String.class
@@ -120,13 +108,13 @@ public class KakaoService {
     }
 
     private String createJwtToken(KakaoUserInfoDto kakaoUserInfo) {
-        User user = userAdapter.createKakaoUser(kakaoUserInfo.getId(), kakaoUserInfo.getNickname(), kakaoUserInfo.getPictureUrl());
+        User user = userService.createKakaoUser(kakaoUserInfo.getId(), kakaoUserInfo.getNickname(), kakaoUserInfo.getPictureUrl());
 
         String accessToken = jwtProvider.createAccessToken(user.getUsername());
         String refreshToken = jwtProvider.createRefreshToken(user.getUsername());
 
         user.setRefreshToken(refreshToken);
-        userAdapter.save(user);
+        userService.save(user);
 
         return accessToken;
     }
