+++ src/main/java/com/sparta/springtrello/domain/user/service/UserService.java
@@ -1,16 +1,16 @@
 package com.sparta.springtrello.domain.user.service;
 
+import com.sparta.springtrello.common.ResponseCodeEnum;
 import com.sparta.springtrello.common.S3Uploader;
 import com.sparta.springtrello.domain.user.dto.*;
 import com.sparta.springtrello.domain.user.entity.User;
 import com.sparta.springtrello.domain.user.entity.UserRoleEnum;
 import com.sparta.springtrello.domain.user.entity.UserStatusEnum;
-import com.sparta.springtrello.domain.user.repository.UserAdapter;
+import com.sparta.springtrello.domain.user.repository.UserRepository;
 import com.sparta.springtrello.exception.custom.common.UploadException;
-import com.sparta.springtrello.exception.custom.user.UserException;
 import com.sparta.springtrello.exception.custom.user.PasswordException;
-import com.sparta.springtrello.common.ResponseCodeEnum;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.sparta.springtrello.exception.custom.user.UserException;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -18,28 +18,52 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class UserService {
 
-    private final UserAdapter userAdapter;
+    private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
-    @Value("${manager-password}")
-    private String managerPassword;
     private final S3Uploader s3Uploader;
 
-    @Autowired
-    public UserService(UserAdapter userAdapter, PasswordEncoder passwordEncoder, S3Uploader s3Uploader) {
-        this.userAdapter = userAdapter;
-        this.passwordEncoder = passwordEncoder;
-        this.s3Uploader = s3Uploader;
+    @Value("${manager-password}")
+    private String managerPassword;
+
+    public User findById(Long id) {
+        return userRepository.findById(id)
+                .orElseThrow(() -> new UserException(ResponseCodeEnum.USER_NOT_FOUND));
+    }
+
+    public User findByUsername(String username) {
+        return userRepository.findByUsername(username)
+                .orElseThrow(() -> new UserException(ResponseCodeEnum.USER_NOT_FOUND));
+    }
+
+    public boolean existsByUsername(String username) {
+        return userRepository.findByUsername(username).isPresent();
+    }
+
+    public User save(User user) {
+        return userRepository.save(user);
+    }
+
+    public User createKakaoUser(Long kakaoId, String nickname, String pictureUrl) {
+        return userRepository.findByKakaoId(kakaoId).orElseGet(() -> {
+            String password = UUID.randomUUID().toString();
+            String encodedPassword = passwordEncoder.encode(password);
+            User newUser = new User(kakaoId, "kakao_" + kakaoId, pictureUrl, encodedPassword);
+            newUser.setNickname(nickname);
+            userRepository.save(newUser);
+            return newUser;
+        });
     }
 
     // 회원가입
     @Transactional
     public void signup(SignupRequestDto requestDto) {
-        // 사용자명이 이미 존재하는지 확인
-        if (userAdapter.existsByUsername(requestDto.getUsername())) {
+        if (existsByUsername(requestDto.getUsername())) {
             throw new UserException(ResponseCodeEnum.USER_ALREADY_EXISTS);
         }
 
@@ -51,7 +75,6 @@ public class UserService {
                 UserStatusEnum.STATUS_NORMAL
         );
 
-        // 사용자 ROLE 확인
         UserRoleEnum role = UserRoleEnum.ROLE_USER;
         if (requestDto.getManagerPassword() != null && !requestDto.getManagerPassword().isEmpty()) {
             if (!managerPassword.equals(requestDto.getManagerPassword())) {
@@ -60,28 +83,25 @@ public class UserService {
             role = UserRoleEnum.ROLE_MANAGER;
         }
         user.setUserRole(role);
-        userAdapter.save(user);
+        save(user);
     }
 
     // 비밀번호 변경
     @Transactional
     public void updatePassword(User loginUser, PasswordUpdateRequestDto requestDto) {
-        User user = userAdapter.findById(loginUser.getId());
+        User user = findById(loginUser.getId());
 
-        // 현재 비밀번호 확인
         if (!passwordEncoder.matches(requestDto.getCurrentPassword(), user.getPassword())) {
             throw new PasswordException(ResponseCodeEnum.PASSWORD_INCORRECT);
         }
 
-        // 새로운 비밀번호가 현재 비밀번호와 다른지 확인
         if (passwordEncoder.matches(requestDto.getNewPassword(), user.getPassword())) {
             throw new PasswordException(ResponseCodeEnum.SAME_AS_OLD_PASSWORD);
         }
 
-        // 비밀번호 업데이트
         String encodedNewPassword = passwordEncoder.encode(requestDto.getNewPassword());
         user.setPassword(encodedNewPassword);
-        userAdapter.save(user);
+        save(user);
     }
 
     // 프로필 업로드
@@ -98,33 +118,32 @@ public class UserService {
                 throw new UploadException(ResponseCodeEnum.UPLOAD_FAILED);
             }
         }
-        userAdapter.save(user);
+        save(user);
     }
 
     // 프로필 조회
     @Transactional(readOnly = true)
     public ProfileResponseDto getProfile(Long userId) {
-        User user = userAdapter.findById(userId);
+        User user = findById(userId);
         return new ProfileResponseDto(user.getNickname(), user.getIntroduce(), user.getPictureUrl());
     }
 
     // 로그아웃
     @Transactional
     public void logout(User user) {
-        user.setRefreshToken(null);  // 리프레시 토큰을 무효화
-        userAdapter.save(user);
+        user.setRefreshToken(null);
+        save(user);
     }
 
     // 회원 탈퇴
     @Transactional
     public void deleteUser(User user, AccountDeleteRequestDto requestDto) {
-        // 비밀번호 확인
         if (!passwordEncoder.matches(requestDto.getPassword(), user.getPassword())) {
             throw new PasswordException(ResponseCodeEnum.PASSWORD_INCORRECT);
         }
 
         user.setUserStatus(UserStatusEnum.STATUS_DELETED);
-        user.setRefreshToken(null); // 리프레시 토큰을 무효화
-        userAdapter.save(user);
+        user.setRefreshToken(null);
+        save(user);
     }
 }
